requirements.txt
numpy==1.18.1
streamlit==0.60.0
matplotlib==3.1.3
keras==2.2.5
tensorflow=1.14
pandas==1.0.1
nsepy==0.8
pandas_datareader==0.8.1
Pillow==7.1.2
scikit_learn==0.23.1